Added Firebase 4.9.0 bianries
Added
- Analytics
- Auth
- DynamicLinks
- Messaging
- RemoteConfig

Those are binaries compiled by Firebase, the only modification is to
CFBundleIdentifier for the frameworks in the Info.plist. Binaries
provided by Google has the same ID `com.firebase.Firebase` which results
in installation error "This app could not be installed at this time."
due to duplicated bundle IDs of embedded frameworks.